Move readjson into readjson folder.
There is only one public method the rest are private.
Therefore there is only one cpp file
