+++ src/readjson/OpenWQ_readjson.cpp
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "OpenWQ_readjson.h"
+#include "OpenWQ_readjson.hpp"
 
 
 /* #################################################
